net/cadence_gem: Fix Update of tx_desc_addr
tx_desc_addr needs to an increment of packet_desc_addr.

Message-Id: <1588278632-23770-4-git-send-email-sai.pavan.boddu@xilinx.com>

+++ hw/net/cadence_gem.c
@@ -1373,7 +1373,7 @@ static void gem_transmit(CadenceGEMState *s)
                 if (tx_desc_get_wrap(desc)) {
                     s->tx_desc_addr[q] = gem_get_tx_queue_base_addr(s, q);
                 } else {
-                    s->tx_desc_addr[q] = packet_desc_addr +
+                    s->tx_desc_addr[q] = (uint32_t)packet_desc_addr +
                                          4 * gem_get_desc_len(s, false);
                 }
                 DB_PRINT("TX descriptor next: 0x%08x\n", s->tx_desc_addr[q]);
